Move terminal interaction modules to top-level src/terminal folder
Relocate the Prompter service and title rendering out of lib into a
dedicated src/terminal directory, mirroring packref's structure. Terminal
interaction is a concern of commands, not core library code, so it now
lives as a sibling of lib and commands.
