+++ src/terminal/title.macro.ts
